+++ jest-bundle-cjs.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    moduleNameMapper: {
+        '~\/src$': '<rootDir>/dist/cjs/index.js',
+    },
+    testEnvironment: 'node',
+};