+++ jest-bundle.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    moduleNameMapper: {
+        '~\/src$': '<rootDir>/dist/index.js',
+    },
+};