+++ src/tx-params/candidacy-declare.js
@@ -1,8 +1,5 @@
-import MinterDeclareCandidacyTxData from 'minterjs-tx/src/tx-data/declare-candidacy';
-import {TX_TYPE_DECLARE_CANDIDACY} from 'minterjs-tx/src/tx-types';
-import {convertToPip} from 'minterjs-util/src/converter';
-import {formatCoin} from 'minterjs-tx/src/helpers';
-import {toBuffer} from 'minterjs-util/src/prefix';
+import {MinterTxDataDeclareCandidacy, TX_TYPE_DECLARE_CANDIDACY, formatCoin} from 'minterjs-tx';
+import {convertToPip, toBuffer} from 'minterjs-util';
 import {toHexString} from '../utils';
 
 /**
@@ -17,7 +14,7 @@ import {toHexString} from '../utils';
  * @return {TxParams}
  */
 export default function DeclareCandidacyTxParams({address, publicKey, commission, coinSymbol, stake, feeCoinSymbol, ...otherParams}) {
-    const txData = new MinterDeclareCandidacyTxData({
+    const txData = new MinterTxDataDeclareCandidacy({
         address: toBuffer(address),
         pubKey: toBuffer(publicKey),
         commission: `0x${toHexString(commission)}`,