+++ src/tx-params/convert-sell-all.js
@@ -1,7 +1,5 @@
-import MinterSellAllTxData from 'minterjs-tx/src/tx-data/sell-all';
-import {formatCoin} from 'minterjs-tx/src/helpers';
-import {TX_TYPE_SELL_ALL_COIN} from 'minterjs-tx/src/tx-types';
-import {convertToPip} from 'minterjs-util/src/converter';
+import {MinterTxDataSellAll, TX_TYPE_SELL_ALL_COIN, formatCoin} from 'minterjs-tx';
+import {convertToPip} from 'minterjs-util';
 
 /**
  * @constructor
@@ -13,7 +11,7 @@ import {convertToPip} from 'minterjs-util/src/converter';
  * @return {TxParams}
  */
 export default function SellAllTxParams({coinFrom, coinTo, minBuyAmount = 0, feeCoinSymbol, ...otherParams}) {
-    const txData = new MinterSellAllTxData({
+    const txData = new MinterTxDataSellAll({
         coinToSell: formatCoin(coinFrom),
         coinToBuy: formatCoin(coinTo),
         minimumValueToBuy: `0x${convertToPip(minBuyAmount, 'hex')}`,