+++ src/tx-params/stake-unbond.js
@@ -1,8 +1,5 @@
-import MinterUnbondTxData from 'minterjs-tx/src/tx-data/unbond';
-import {toBuffer} from 'minterjs-util/src/prefix';
-import {formatCoin} from 'minterjs-tx/src/helpers';
-import {convertToPip} from 'minterjs-util/src/converter';
-import {TX_TYPE_UNBOND} from 'minterjs-tx/src/tx-types';
+import {MinterTxDataUnbond, TX_TYPE_UNBOND, formatCoin} from 'minterjs-tx';
+import {toBuffer, convertToPip} from 'minterjs-util';
 
 /**
  * @constructor
@@ -14,7 +11,7 @@ import {TX_TYPE_UNBOND} from 'minterjs-tx/src/tx-types';
  * @return {TxParams}
  */
 export default function UnbondTxParams({publicKey, coinSymbol, stake, feeCoinSymbol, ...otherParams}) {
-    const txData = new MinterUnbondTxData({
+    const txData = new MinterTxDataUnbond({
         pubKey: toBuffer(publicKey),
         coin: formatCoin(coinSymbol),
         stake: `0x${convertToPip(stake, 'hex')}`,