+++ test/tx-params/create-multisig.test.js
@@ -1,5 +1,5 @@
 import {Buffer} from 'safe-buffer';
-import {TX_TYPE_CREATE_MULTISIG} from 'minterjs-tx/src/tx-types';
+import {TX_TYPE_CREATE_MULTISIG} from 'minterjs-tx';
 import {CreateMultisigTxParams} from '~/src';