+++ test/tx-params/multisend.test.js
@@ -1,5 +1,5 @@
 import {Buffer} from 'safe-buffer';
-import {TX_TYPE_MULTISEND} from 'minterjs-tx/src/tx-types';
+import {TX_TYPE_MULTISEND} from 'minterjs-tx';
 import {MultisendTxParams} from '~/src';